Huge refactor:
Focus on SQLite3 implementation, abandon complicated design for adopting MySQL.
Binding is pushed to be done in SQLite3 instead of evaluating in JavaScript.
Implement native Stmt, which is capable of running multiple SQLs as a group.
Change the way of handing schema change.
Change the way of development: switch back to test-first approach.